+++ WegliKit/Sources/Helper/Array+Additions.swift
@@ -1,0 +1,11 @@
+// Created for weg-li in 2021.
+
+import Foundation
+
+public extension Array {
+  func chunked(into size: Int) -> [[Element]] {
+    stride(from: 0, to: count, by: size).map {
+      Array(self[$0..<Swift.min($0 + size, count)])
+    }
+  }
+}
