+++ WegliKit/Sources/Helper/CLLocationCoordinate2D+Additions.swift
@@ -1,0 +1,34 @@
+// Created for weg-li in 2021.
+
+import CoreLocation
+
+extension CLLocationCoordinate2D: Equatable {
+  public static func == (lhs: CLLocationCoordinate2D, rhs: CLLocationCoordinate2D) -> Bool {
+    lhs.latitude == rhs.latitude && lhs.longitude == rhs.longitude
+  }
+}
+
+public extension CLLocationCoordinate2D {
+  static let zero = CLLocationCoordinate2D(latitude: 0, longitude: 0)
+}
+
+extension CLLocationCoordinate2D: Codable {
+  public func encode(to encoder: Encoder) throws {
+    var container = encoder.unkeyedContainer()
+    try container.encode(longitude)
+    try container.encode(latitude)
+  }
+  
+  public init(from decoder: Decoder) throws {
+    var container = try decoder.unkeyedContainer()
+    let longitude = try container.decode(CLLocationDegrees.self)
+    let latitude = try container.decode(CLLocationDegrees.self)
+    self.init(latitude: latitude, longitude: longitude)
+  }
+}
+
+public extension CLLocation {
+  convenience init(from coordinate: CLLocationCoordinate2D) {
+    self.init(latitude: coordinate.latitude, longitude: coordinate.longitude)
+  }
+}
