+++ WegliKit/Sources/Helper/String+Additions.swift
@@ -1,0 +1,9 @@
+// Created for weg-li in 2021.
+
+import Foundation
+
+public extension String {
+  var isNumeric: Bool {
+    CharacterSet.decimalDigits.isSuperset(of: CharacterSet(charactersIn: self))
+  }
+}
