--- weg-li/Utils/Extensions/Array+Additions.swift
@@ -1,19 +1,0 @@
-// Created for weg-li in 2021.
-
-import Foundation
-
-extension Array {
-    func chunked(into size: Int) -> [[Element]] {
-        stride(from: 0, to: count, by: size).map {
-            Array(self[$0..<Swift.min($0 + size, count)])
-        }
-    }
-}
-
-extension Array where Element == District {
-    static let all = Bundle.main.decode(
-        [District].self, from: "districts.json",
-        dateDecodingStrategy: .iso8601,
-        keyDecodingStrategy: .convertFromSnakeCase
-    )
-}
