--- weg-li/Utils/Extensions/Bundle+Additions.swift
@@ -1,48 +1,0 @@
-// Created for weg-li in 2021.
-
-import Foundation
-
-extension Bundle {
-    func decode<T: Decodable>(
-        _ type: T.Type,
-        from file: String,
-        dateDecodingStrategy: JSONDecoder.DateDecodingStrategy = .deferredToDate,
-        keyDecodingStrategy: JSONDecoder.KeyDecodingStrategy = .useDefaultKeys
-    ) -> T {
-        guard let url = self.url(forResource: file, withExtension: nil) else {
-            fatalError("Failed to locate \(file) in bundle.")
-        }
-
-        guard let data = try? Data(contentsOf: url) else {
-            fatalError("Failed to load \(file) from bundle.")
-        }
-
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = dateDecodingStrategy
-        decoder.keyDecodingStrategy = keyDecodingStrategy
-
-        do {
-            return try decoder.decode(T.self, from: data)
-        } catch let DecodingError.keyNotFound(key, context) {
-            fatalError("Failed to decode \(file) from bundle due to missing key '\(key.stringValue)' not found – \(context.debugDescription)")
-        } catch let DecodingError.typeMismatch(_, context) {
-            fatalError("Failed to decode \(file) from bundle due to type mismatch – \(context.debugDescription)")
-        } catch let DecodingError.valueNotFound(type, context) {
-            fatalError("Failed to decode \(file) from bundle due to missing \(type) value – \(context.debugDescription)")
-        } catch DecodingError.dataCorrupted(_) {
-            fatalError("Failed to decode \(file) from bundle because it appears to be invalid JSON")
-        } catch {
-            fatalError("Failed to decode \(file) from bundle: \(error.localizedDescription)")
-        }
-    }
-}
-
-extension Bundle {
-    var versionNumber: String {
-        infoDictionary?["CFBundleShortVersionString"] as? String ?? ""
-    }
-
-    var buildNumber: String {
-        infoDictionary?["CFBundleVersion"] as? String ?? ""
-    }
-}
