--- weg-li/Utils/Extensions/Color+Additions.swift
@@ -1,7 +1,0 @@
-// Created for weg-li in 2021.
-
-import SwiftUI
-
-extension Color {
-    static let wegliBlue = Color("wegliBlue")
-}
